+++ src/ex_raw.c
@@ -149,7 +149,7 @@ void draw_raw_style_end_cap (int ncp,		/* number of contour points */
 
 #ifdef OPENGL_10
    /* malloc the @#$%^&* array that OpenGL wants ! */
-   pts = (double *) malloc (3*ncp*sizeof(double));
+   pts = (double *) malloc (3*(size_t)ncp*sizeof(double));
    tobj = gluNewTess ();
    gluTessCallback (tobj, GLU_BEGIN, glBegin);
    gluTessCallback (tobj, GLU_VERTEX, glVertex3dv);
@@ -788,7 +788,7 @@ void extrusion_raw_join (int ncp,		/* number of contour points */
 
    /* alloc loop arrays if needed */
    if (! no_xform) {
-      mem_anchor = malloc (4 * ncp * 3 * sizeof(gleDouble));
+      mem_anchor = malloc (4 * (size_t)ncp * 3 * sizeof(gleDouble));
       front_loop = (gleDouble *) mem_anchor;
       back_loop = front_loop + 3*ncp;
       front_norm = back_loop + 3*ncp;
